fix: extract shared artifact download utility to eliminate duplication
